+++ test/integration/016_macro_tests/macros/my_macros.sql
@@ -15,8 +15,23 @@
 {% endmacro %}
 
 
-{# there is no no default__dispatch_to_nowhere! #}
+{# there is no default__dispatch_to_nowhere! #}
 {% macro dispatch_to_nowhere() %}
 	{% set macro = adapter.dispatch('dispatch_to_nowhere') %}
 	{{ macro() }}
+{% endmacro %}
+
+
+{% macro dispatch_to_parent() %}
+	{% set macro = adapter.dispatch('dispatch_to_parent') %}
+	{{ macro() }}
+{% endmacro %}
+
+{% macro default__dispatch_to_parent() %}
+	{% set msg = 'No default implementation of dispatch_to_parent' %}
+    {{ exceptions.raise_compiler_error(msg) }}
+{% endmacro %}
+
+{% macro postgres__dispatch_to_parent() %}
+	{{ return('') }}
 {% endmacro %}