+++ test/integration/016_macro_tests/test_macros.py
@@ -95,6 +95,20 @@ class TestAdapterMacroNoDestination(DBTIntegrationTest):
         assert "In dispatch: No macro named 'dispatch_to_nowhere' found" in str(exc.value)
 
 
+class TestDispatchMacroUseParent(DBTIntegrationTest):
+    @property
+    def schema(self):
+        return "test_macros_016"
+
+    @property
+    def models(self):
+        return "dispatch-inheritance-models"
+
+    @use_profile('redshift')
+    def test_redshift_inherited_macro(self):
+        self.run_dbt(['run'])
+
+
 class TestMacroOverrideBuiltin(DBTIntegrationTest):
     @property
     def schema(self):